refactor: Move commen code into `apiBase.ts`
